requirements.txt
asgiref==3.8.1
coverage==7.6.12
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
freezegun==1.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.4
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
